de-duplicate archive unpacking
